changed data type of all blood sugar related fields or edit texts

+++ app/src/main/java/de/dmate/marvin/dmate/activities/NewEntryActivity.java
@@ -86,7 +86,7 @@ public class NewEntryActivity extends AppCompatActivity
                 if (requestCode == 1) {
                     newEntry.setTimestamp(new Timestamp(calendar.getTimeInMillis()));
 
-                    if (!ETbloodsugar.getText().toString().equals("")) newEntry.setBloodSugar(Integer.parseInt(ETbloodsugar.getText().toString()));
+                    if (!ETbloodsugar.getText().toString().equals("")) newEntry.setBloodSugar(Float.parseFloat(ETbloodsugar.getText().toString()));
                     else newEntry.setBloodSugar(null);
 
                     if (!ETbreadunit.getText().toString().equals("")) newEntry.setBreadUnit(Float.parseFloat(ETbreadunit.getText().toString()));
@@ -106,7 +106,7 @@ public class NewEntryActivity extends AppCompatActivity
 
                 if (requestCode == 2) {
                     currentEntry.setTimestamp(new Timestamp(calendar.getTimeInMillis()));
-                    if (!ETbloodsugar.getText().toString().equals("")) currentEntry.setBloodSugar(Integer.parseInt(ETbloodsugar.getText().toString()));
+                    if (!ETbloodsugar.getText().toString().equals("")) currentEntry.setBloodSugar(Float.parseFloat(ETbloodsugar.getText().toString()));
                     if (!ETbreadunit.getText().toString().equals("")) currentEntry.setBreadUnit(Float.parseFloat(ETbreadunit.getText().toString()));
                     if (!ETbolus.getText().toString().equals("")) currentEntry.setBolus(Float.parseFloat(ETbolus.getText().toString()));
                     if (!ETbasal.getText().toString().equals("")) currentEntry.setBasal(Float.parseFloat(ETbasal.getText().toString()));
@@ -188,7 +188,7 @@ public class NewEntryActivity extends AppCompatActivity
             dateButton.setText(Helper.formatMillisToDateString(currentEntry.getTimestamp().getTime()));
             timeButton.setText(Helper.formatMillisToTimeString(currentEntry.getTimestamp().getTime()));
 
-            Integer bloodsugar = currentEntry.getBloodSugar();
+            Float bloodsugar = currentEntry.getBloodSugar();
             if (bloodsugar != null) ETbloodsugar.setText(bloodsugar.toString());
 
             Float breadunit = currentEntry.getBreadUnit();
